requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.84
botocore==1.34.84
build==1.2.1
cachetools==5.3.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==40.0.2
cycler==0.12.1
Cython==0.29.37
dataclasses-json==0.6.3
datasets==2.19.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.8
distro==1.8.0
dnspython==2.6.1
docker==7.0.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.2
fastjsonschema==2.19.1
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.4.0
ipywidgets==8.1.2
isoduration==20.11.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5
langsmith==0.0.83
Levenshtein==0.21.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.10.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.1
pathos==0.3.2
Pillow==9.5.0
platformdirs==4.2.1
posthog==3.5.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0
qtconsole==5.5.1
QtPy==2.4.1
rank-bm25==0.2.2
rapidfuzz==3.1.1
referencing==0.34.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.3.1
sagemaker==2.219.0
schema==0.7.7
scikit-learn==1.2.2
scipy==1.11.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
smdebug-rulesconfig==1.0.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.37.2
sympy==1.12
tblib==3.0.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.1.0
tiktoken==0.5.2
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
tzdata==2023.3
ujson==5.9.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2